requirements.txt
black >= 24.2.0
Django >= 4.2.0, < 5.0.0
django-ninja >= 1.1.0, < 2.0.0
pytest >= 8.0.1, < 9.0.0
pytest-django >= 4.8.0, < 5.0.0
requests >= 2.31.0, < 3.0.0